[Silver III] Title: 팰린드롬 만들기, Time: 64 ms, Memory: 11484 KB -BaekjoonHub

+++ 백준/Silver/1213. 팰린드롬 만들기/팰린드롬 만들기.java
@@ -6,14 +6,14 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         char[] input = br.readLine().toCharArray();
-        int[] count = new int[26];
-        int length = input.length;
+        byte[] count = new byte[26];
+        byte length = (byte) input.length;
 
         for (char ch : input) {
             count[ch - 'A']++;
         }
 
-        int odd = 0;
+        byte odd = 0;
         char oddChar = 0;
         for (int i = 0; i < 26; i++) {
             if ((count[i] & 1) == 1) {
@@ -27,8 +27,8 @@ public class Main {
         }
 
         char[] result = new char[length];
-        int left = 0, right = length - 1;
-        
+        byte left = 0, right = (byte) (length - 1);
+
         for (int i = 0; i < 26; i++) {
             while (count[i] > 1) {
                 result[left++] = (char) (i + 'A');
